docker-compose.yml & Dockerfile cleanup.
docker-compose already has the host current dir mapped to /code so there
is no need for COPY . /code/ to copy the code into the dockerfile.

WORKDIR /code is set in the dockerfile   so docker-compose can just use
 ./manage.py to run the server.

Changed the README.md to use docker-compose build vs docker build . as
the build parameters are contained in the docker-compose.yml  file.