Update _base.html
This pull request addresses a styling issue in the navbar menu. The items that should be on the right side of the menu were not rendering correctly due to the <ul> tag closing within the {% else %} block. Additionally, the styling for the form with the class form-dflex me-auto was adjusted to ensure proper alignment.

Changes:

Added </ul> to close the <ul> tag under the {% else %} block.
Adjusted the styling for the form with the class form-dflex me-auto for proper alignment.
Why This Change Was Made:
To ensure proper rendering of the navbar menu and form alignment, it was necessary to close the <ul> tag within the {% else %} block and adjust the form styling accordingly. These fixes ensure that menu items are displayed correctly and that the form aligns properly.

Technical Explanation:
Closing the <ul> tag within the {% else %} block ensures correct HTML structuring for the navbar menu, allowing items on both the right and left sides to render correctly. Additionally, adjusting the styling for the form with the class form-dflex me-auto ensures proper alignment according to the Bootstrap layout. These changes contribute to improved visual consistency and user experience.