Update README.md docker commands
Changed 'docker-compose' to 'docker compose' in the install instructions because latest docker version have update from python to go, therefore 'docker-compose' is now outdated and obsolete in future versions of docker. So users using the latest versions of docker will not be able to use 'docker-compose' to but MUST use 'docker compose' instead. 

I had to read up on stackO how to install docker which led me to the official docs to uninstall and reinstall the official docker. Then i couldnt check the version when i used the dashes so i went back to stackO and found out the commands have changed.